changed order of factors in shading calculation, acording to textbook

+++ sphere.cpp
@@ -48,7 +48,7 @@ zie ook
   Vector distance = ray.O - position;
   Vector direction = ray.D.normalized();
 
-  int discriminant = direction.dot(distance) * direction.dot(distance) - distance.dot(distance) + r*r;
+  double discriminant = direction.dot(distance) * direction.dot(distance) - distance.dot(distance) + r*r;
 
 
   if(discriminant < 0) return Hit::NO_HIT();
@@ -56,7 +56,7 @@ zie ook
   double intersect1 = -(direction.dot(distance)) + sqrt(discriminant);
   double intersect2 = -(direction.dot(distance)) - sqrt(discriminant);
 
-  double t = (distance.length() < r) ? max(intersect1,intersect2) : min(intersect1,intersect2);
+double t = (distance.length() < r) ? max(intersect1,intersect2) : min(intersect1,intersect2);
 
    
     /****************************************************
